pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "localflow"
version = "0.1.0"
description = "Fully local Wispr Flow-style dictation assistant."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "faster-whisper>=1.0.3",
  "numpy>=1.26",
  "pynput>=1.7.7",
  "pyautogui>=0.9.54",
  "pyperclip>=1.9.0",
  "sounddevice>=0.4.7",
]

[project.optional-dependencies]
rewrite = ["llama-cpp-python>=0.2.90"]

[project.scripts]
localflow = "localflow.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
